pyproject.toml
[tool.poetry]
name = "piri-web"
version = "0.1.0"
description = "Piri Json to Json API Server"
authors = ["Thomas Borgen <thomas@borgenit.no>"]

[tool.poetry.dependencies]
python = "^3.8"
piri = "^1.0.2"
falcon = "^2.0.0"
simplejson = "^3.17.2"
attrs = "^20.1.0"

[tool.poetry.dev-dependencies]
# Tests
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
# Code quality
wemake-python-styleguide = "^0.14.1"
mypy = "^0.782"
safety = "^1.9.0"
# docs
mkdocs-material = "^5.5.8"
# local dev
gunicorn = "^20.0.4"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
